# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Holiday Plans - Docs",
        "description": "API documentation for the <b>Holiday Plans</b> project",
        "contact": {
            "email": "yuricaparelliofc@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api",
            "description": "API Url"
        }
    ],
    "paths": {
...
        "/holiday-plans/": {
            "get": {
                "tags": [
                    "HolidayPlans"
                ],
                "summary": "Get all registered holiday plans",
                "description": "Endpoint to retrieve all holiday plans.",
                "operationId": "getHolidayPlans",
                "responses": {
                    "200": {
                        "description": "Success, holiday plans found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": "20"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Travel to Lisbon"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Travel from the best city of europe"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-14"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Lisbon"
                                                },
                                                "participants": {
                                                    "type": "integer",
                                                    "example": "3"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Holiday plan doesn't exists."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid token."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal error, contact an administrator."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/holiday-plan": {
            "post": {
                "tags": [
                    "HolidayPlans"
                ],
                "summary": "Create a holiday plan",
                "description": "Endpoint to create a holiday plan.",
                "operationId": "createHolidayPlan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Travel to Lisbon"
                                    },
                                    "descrtiption": {
                                        "type": "string",
                                        "example": "Travel from the best city of europe"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-03-14"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Lisbon"
                                    },
                                    "participants": {
                                        "type": "integer",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success, holiday plan created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Travel to Lisbon"
                                                },
                                                "descrtiption": {
                                                    "type": "string",
                                                    "example": "Travel from the best city of europe"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-14"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Lisbon"
                                                },
                                                "participants": {
                                                    "type": "integer",
                                                    "example": "3"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": "20"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-14T01:57:56.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-14T01:57:56.000000Z"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal error, contact an administrator."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/holiday-plans/{id}/pdf": {
            "get": {
                "tags": [
                    "HolidayPlans"
                ],
                "summary": "Generate Pdf of a holiday plan by id",
                "description": "Endpoint to generate a Holiday plan pdf.",
                "operationId": "generatePdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Holiday plan ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, pdf generated.",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "description": "PDF file",
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Holiday plan doesn't exists."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient permission.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "You don't have permission to update or delete this Holiday plan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid token."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal error, contact an administrator."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Insert the token in the format: <b>Bearer {token}</b>",
                "name": "Authorization",
                "in": "header",
                "scheme": "sanctum"
            }
        }
    }
}